+++ CHANGELOG.md
@@ -1,3 +1,21 @@
+## [1.0.0](https://github.com/shivam091/postgresql_adapter_extensions/compare/v0.1.0...v1.0.0) - 2025-03-12
+
+### What's new
+
+- Added `create_sequence` method
+
+  Allows creating PostgreSQL sequences with customizable options such as start value, increment step, min/max values, caching, cycling, and ownership.
+
+- Added `alter_sequence` method
+
+  Enables modifying existing PostgreSQL sequences, supporting changes to increment steps, restart values, min/max limits, caching, cycling behavior, and ownership.
+
+- Added `drop_sequence` method
+
+  Provides functionality to remove a PostgreSQL sequence with optional IF EXISTS and CASCADE/RESTRICT behaviors to manage dependencies.
+
+----------
+
 ## 0.1.0 - 2025-03-11
 
 ### Initial release